+++ tests/common/macsec/macsec_platform_helper.py
