+++ tests/macsec/conftest.py
@@ -1,6 +1,6 @@
 import pytest
 
-from .macsec_helper import check_appl_db
+from tests.common.macsec.macsec_helper import check_appl_db
 from tests.common.utilities import wait_until
