+++ tests/macsec/test_deployment.py
@@ -3,7 +3,7 @@ import logging
 
 from tests.common.utilities import wait_until
 from tests.common import config_reload
-from .macsec_helper import check_appl_db
+from tests.common.macsec.macsec_helper import check_appl_db
 logger = logging.getLogger(__name__)
 
 pytestmark = [
