+++ tests/macsec/test_interop_wan_isis.py
@@ -3,9 +3,9 @@ import logging
 
 from tests.common.utilities import wait_until
 from tests.common.helpers.assertions import pytest_assert
-from .macsec_platform_helper import get_portchannel
-from .macsec_platform_helper import find_portchannel_from_member
-from .macsec_config_helper import enable_macsec_port, disable_macsec_port
+from tests.common.macsec.macsec_platform_helper import get_portchannel
+from tests.common.macsec.macsec_platform_helper import find_portchannel_from_member
+from tests.common.macsec.macsec_config_helper import enable_macsec_port, disable_macsec_port
 
 
 logger = logging.getLogger(__name__)
